[Rename]#2 Rename main to root
