+++ app/src/main/java/com/gallery/kakaogallery/presentation/ui/root/FragmentHandler.kt
@@ -1,4 +1,4 @@
-package com.gallery.kakaogallery.presentation.ui.main
+package com.gallery.kakaogallery.presentation.ui.root
 
 import com.gallery.kakaogallery.presentation.ui.comp.HeaderComp
