+++ app/src/main/java/com/gallery/kakaogallery/presentation/ui/root/RootViewModel.kt
@@ -1,8 +1,8 @@
-package com.gallery.kakaogallery.presentation.ui.main
+package com.gallery.kakaogallery.presentation.ui.root
 
 import com.gallery.kakaogallery.presentation.ui.base.BaseViewModel
 
-class MainViewModel : BaseViewModel() {
+class RootViewModel : BaseViewModel() {
 
     var currentPage : Int = 0
 }
